requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
sqlalchemy==2.0.25
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic-settings==2.2.1